pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Harry5174", email = "harisjaved010@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.7",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pypdf>=6.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
